requirements.txt
bcrypt==4.0.1
customtkinter==5.2.0
matplotlib==3.7.2
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
python-dotenv==1.0.0
reportlab==4.0.4 